Volume sliders work
Solution: dart:async periodic timer that calls an update function.
Update function is a setstate that converts a collection of objects in non-widget form to List<Widget>, which is then used as the content on the screen.

This is necessary when you need to call setState on something like a slider or button that is outside of build scope due to being a returned structure