show audio loading progress
when a AudioTile is selected, the program fetches the audio file in real time. the progress of this fetch is now being tracked by `XMLHttpRequest.onprogress`, such that the volume slider is temporarily overriden to communicate the loading progress. after the load is complete, the volume slider returns to it's previous state